requirements.txt
anthropic>=0.40.0
python-chess>=1.9.4
requests>=2.31.0
python-dotenv>=1.0.0
